requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
python-multipart==0.0.9
python-docx==1.1.0
PyPDF2==3.0.1
scikit-learn==1.4.2
numpy==1.26.4
google-generativeai==0.5.2
pydantic==1.10.14
